process_centric_services/movie_match_service/requirements.txt
fastapi
requests
uvicorn
pydantic
passlib[bcrypt]
pyjwt
typing
python-multipart
pytz